requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
